requirements.txt
# Core
python-dotenv>=1.0.1
chromadb>=0.5.4
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5


# LLMs
google-generativeai>=0.7.2
sentence-transformers>=3.0.1


# Text extraction
pypdf>=4.2.0
python-docx>=1.1.0
python-pptx>=0.6.23
markdown>=3.6
beautifulsoup4>=4.12


# OCR / media
pytesseract>=0.3.10
Pillow>=10.4.0
opencv-python>=4.10.0
whisper>=1.1.10
moviepy>=1.0.3
pytube>=15.0.0


# App
streamlit>=1.38